面试预约相关

+++ src/main/java/com/tranzvision/gd/TZInterviewArrangementBundle/service/impl/TzInterviewSetStudentImpl.java
@@ -306,8 +306,11 @@ public class TzInterviewSetStudentImpl extends FrameworkImpl {
 		String strRtn = "";
 		Map<String, Object> rtnMap = new HashMap<String, Object>();
 		rtnMap.put("success", "");
+		rtnMap.put("notAllow", "");
 
 		JacksonUtil jacksonUtil = new JacksonUtil();
+		String notAllowStu = "";
+		int count = 0;
 		try {
 			for (int i = 0; i < actData.length; i++) {
 				String actForm = actData[i];
@@ -316,18 +319,38 @@ public class TzInterviewSetStudentImpl extends FrameworkImpl {
 				String classID = jacksonUtil.getString("classID");
 				String batchID = jacksonUtil.getString("batchID");
 				Long appId = Long.valueOf(jacksonUtil.getString("appId"));
+				
+				//如果考生被面试评委抽取，则不允许删除
+				String sql = "select count(1) from PS_TZ_MP_PW_KS_TBL where TZ_CLASS_ID=? and TZ_APPLY_PC_ID=? and TZ_APP_INS_ID=?";
+				int pwCount = sqlQuery.queryForObject(sql, new Object[]{ classID, batchID, appId }, "int");
+				
+				if(pwCount > 0){
+					if(count < 10){
+						String stuName = jacksonUtil.getString("stuName");
+						if(notAllowStu == ""){
+							notAllowStu = stuName;
+						}else{
+							notAllowStu = "," + stuName;
+						}
+					}
+					count ++;
+				}else{
+					PsTzMsPskshTblKey psTzMspsKshTblKey = new PsTzMsPskshTblKey();
+					psTzMspsKshTblKey.setTzClassId(classID);
+					psTzMspsKshTblKey.setTzApplyPcId(batchID);
+					psTzMspsKshTblKey.setTzAppInsId(appId);
 
-				PsTzMsPskshTblKey psTzMspsKshTblKey = new PsTzMsPskshTblKey();
-				psTzMspsKshTblKey.setTzClassId(classID);
-				psTzMspsKshTblKey.setTzApplyPcId(batchID);
-				psTzMspsKshTblKey.setTzAppInsId(appId);
-
-				PsTzMsPskshTbl psTzMspsKshTbl = psTzMspsKshTblMapper.selectByPrimaryKey(psTzMspsKshTblKey);
-				if (psTzMspsKshTbl != null) {
-					psTzMspsKshTblMapper.deleteByPrimaryKey(psTzMspsKshTblKey);
+					PsTzMsPskshTbl psTzMspsKshTbl = psTzMspsKshTblMapper.selectByPrimaryKey(psTzMspsKshTblKey);
+					if (psTzMspsKshTbl != null) {
+						psTzMspsKshTblMapper.deleteByPrimaryKey(psTzMspsKshTblKey);
+					}
 				}
 			}
+			if(count >= 10){
+				notAllowStu = notAllowStu + ",...";
+			}
 			rtnMap.replace("success", "success");
+			rtnMap.replace("notAllow", notAllowStu);
 		} catch (Exception e) {
 			e.printStackTrace();
 			errMsg[0] = "1";
